// generated/src/app/api-models/cieb-contact.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebContact {
  updateProcess: string;
  updateUser: string;
  updateDatetime: Date;
  insertProcess: string;
  insertUser: string;
  insertDatetime: Date;
  termDate: Date;
  effDate: Date;
  phoneExt: string;
  phoneNum: string;
  phonePrefix: string;
  emailAddress: string;
  lastName: string;
  firstName: string;
  contactCode: string;
  seqEntityId: number;
  seqContactId: number;
}